requirements.txt
requests
requests[socks]
stem
beautifulsoup4
SQLAlchemy
pyyaml
python-dateutil
lxml
dateparser
websocket-client
python-socks
